feat(uploads): modal upload + search + rich cards on sources page (#459)
Moves manual upload UX on /podcasts from an inline drop zone and a minimal
card into the same treatment as podcast episodes:

- Replaces the inline drag-and-drop with an "Upload audio" button that
  opens the existing AudioUpload component inside a closable dialog.
- Adds a search input that filters manual uploads by title and
  description (case-insensitive).
- Renders each upload with the same tag strip used on podcast episode
  pages (status, date, duration, language, provider, transcribe/diarize
  durations, Fireworks cost, speaker names, processing progress, and
  failure details), by extracting the card body of EpisodesList into a
  shared EpisodeCard component.
- Keeps the per-upload delete affordance from #454, now in the card's
  bottom-right corner with a confirmation prompt.

Closes #459
